+++ chalice/awsclient.py
@@ -672,8 +672,21 @@ class TypedAWSClient(object):
         # type: (str, bool) -> Iterator[Dict[str, Any]]
         logs = self._client('logs')
         paginator = logs.get_paginator('filter_log_events')
-        for page in paginator.paginate(logGroupName=log_group_name,
-                                       interleaved=True):
+        pages = paginator.paginate(logGroupName=log_group_name,
+                                   interleaved=True)
+        try:
+            for log_message in self._iter_log_messages(pages):
+                yield log_message
+        except logs.exceptions.ResourceNotFoundException:
+            # If the lambda function exists but has not been invoked yet,
+            # it's possible that the log group does not exist and we'll get
+            # a ResourceNotFoundException.  If this happens we return instead
+            # of propagating an exception back to the user.
+            pass
+
+    def _iter_log_messages(self, pages):
+        # type: (Iterator[Dict[str, Any]]) -> Iterator[Dict[str, Any]]
+        for page in pages:
             events = page['events']
             for event in events:
                 # timestamp is modeled as a 'long', so we'll